Remove useMenuRoot hook
